Move getFeaturedEntry to shared Utils library
